Remove extra apt commands and adjust package versions
This change adds logic to remove run commands associated with the detected package manager. Specifically for the apt related commands
`apt-add-repository` and `add-apt-repository`. These commands are
used to add additional repositories from which to install packages.
This packages will be migrated to `apk` so these commands should
be removed.

Relateddto the above apt commands, a empty mapping for `software-properties-common`
is added which will effectively remove this package from the converted
package installs.

Finally, in order to identiffy and perform package mappings an attempt
is made to parse the package name and version information from the
install argument. Previously if a package included a version specifier
it would not be mapped.

When converting a package install when a version is specified some
liberties are taken to align with Chainguard recommended best practices.
Namely release specifiers are dropped and fuzzy matching is forced.

Fixes: #71, #78